refactor(core): extract TTS text chunking into execution::text_chunking
Move the center-break TTS chunking algorithm off TemplateExecutor's impl
block — chunk_text_for_tts, center_break_split, find_nearest,
find_nearest_break_word, migrate_trailing_break_words and the BREAK_WORDS
const — into a new execution::text_chunking module as free functions
(M-REGULAR-FN), with module docs and the 13 chunking tests relocated.
execute_tts_chunked now delegates to text_chunking::chunk_text_for_tts.
No behavior change; executor.rs drops ~400 lines.
